rename the file and making the changes related to comments
